requirements.txt
pip==25.1.1
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.1.0
pydantic==2.6.3
groq==0.4.2
transformers==4.51.3
